Move channel logging into instrumentation
