+++ test/phoenix/logger_test.exs
@@ -1,4 +1,4 @@
-defmodule Phoenix.Controller.LoggerTest do
+defmodule Phoenix.LoggerTest do
   use ExUnit.Case, async: true
   use RouterHelper
   import ExUnit.CaptureLog
@@ -35,7 +35,7 @@ defmodule Phoenix.Controller.LoggerTest do
       |> put_private(:phoenix_pipelines, [:browser])
       |> action
     end
-    assert output =~ "[debug] Processing by Phoenix.Controller.LoggerTest.LoggerController.index/2"
+    assert output =~ "[debug] Processing by Phoenix.LoggerTest.LoggerController.index/2"
     assert output =~ "Parameters: %{\"foo\" => \"bar\", \"format\" => \"html\"}"
     assert output =~ "Pipelines: [:browser]"
   end
@@ -70,6 +70,34 @@ defmodule Phoenix.Controller.LoggerTest do
       |> action
     end
     assert output =~ "Parameters: [UNFETCHED]"
+  end
+
+  test "filter_values" do
+    assert Phoenix.Logger.filter_values(%{"foo" => "bar", "password" => "should_not_show"}, ["password"]) ==
+           %{"foo" => "bar", "password" => "[FILTERED]"}
+  end
+
+  test "filter_values when a map has secret key" do
+    assert Phoenix.Logger.filter_values(%{"foo" => "bar", "map" => %{"password" => "should_not_show"}}, ["password"]) ==
+           %{"foo" => "bar", "map" => %{"password" => "[FILTERED]"}}
+  end
+
+  test "filter_values when a list has a map with secret" do
+    assert Phoenix.Logger.filter_values(%{"foo" => "bar", "list" => [%{"password" => "should_not_show"}]}, ["password"]) ==
+           %{"foo" => "bar", "list" => [%{"password" => "[FILTERED]"}]}
+  end
+
+  test "filter_values does not filter structs" do
+    assert Phoenix.Logger.filter_values(%{"foo" => "bar", "file" => %Plug.Upload{}}, ["password"]) ==
+           %{"foo" => "bar", "file" => %Plug.Upload{}}
+
+    assert Phoenix.Logger.filter_values(%{"foo" => "bar", "file" => %{__struct__: "s"}}, ["password"]) ==
+           %{"foo" => "bar", "file" => %{:__struct__ => "s"}}
+  end
+
+  test "filter_values does not fail on atomic keys" do
+    assert Phoenix.Logger.filter_values(%{:foo => "bar", "password" => "should_not_show"}, ["password"]) ==
+           %{:foo => "bar", "password" => "[FILTERED]"}
   end
 
   defp action(conn) do
