+++ test/phoenix/test/channel_test.exs
@@ -177,6 +177,7 @@ defmodule Phoenix.Test.ChannelTest do
 
   ## join
 
+  @tag :capture_log
   test "join/3 with success" do
     assert {:ok, socket, client} = join(socket("id", original: :assign), Channel, "foo:socket")
     assert socket.channel == Channel
@@ -193,11 +194,13 @@ defmodule Phoenix.Test.ChannelTest do
     assert client.channel_pid in links
   end
 
+  @tag :capture_log
   test "join/3 with error reply" do
     assert {:error, %{reason: "mybad"}} =
              join(socket(), Channel, "foo:error", %{"error" => "mybad"})
   end
 
+  @tag :capture_log
   test "join/3 with crash" do
     Process.flag(:trap_exit, true)
     Logger.disable(self())
@@ -207,6 +210,7 @@ defmodule Phoenix.Test.ChannelTest do
 
   ## handle_in
 
+  @tag :capture_log
   test "pushes and receives pushed messages" do
     {:ok, _, socket} = join(socket(), Channel, "foo:ok")
     ref = push socket, "noreply", %{"req" => "foo"}
@@ -214,6 +218,7 @@ defmodule Phoenix.Test.ChannelTest do
     refute_reply ref, _status
   end
 
+  @tag :capture_log
   test "pushes and receives replies" do
     {:ok, _, socket} = join(socket(), Channel, "foo:ok")
 
@@ -225,6 +230,7 @@ defmodule Phoenix.Test.ChannelTest do
     assert_reply ref, :ok, %{"resp" => "foo"}
   end
 
+  @tag :capture_log
   test "receives async replies" do
     {:ok, _, socket} = join(socket(), Channel, "foo:ok")
 
@@ -232,6 +238,7 @@ defmodule Phoenix.Test.ChannelTest do
     assert_reply ref, :ok, %{"async_resp" => "foo"}
   end
 
+  @tag :capture_log
   test "pushes on stop" do
     Process.flag(:trap_exit, true)
     {:ok, _, socket} = join(socket(), Channel, "foo:ok")
@@ -245,6 +252,7 @@ defmodule Phoenix.Test.ChannelTest do
     push socket, "stop", %{"reason" => :normal}
   end
 
+  @tag :capture_log
   test "pushes and receives replies on stop" do
     Process.flag(:trap_exit, true)
 
@@ -263,6 +271,7 @@ defmodule Phoenix.Test.ChannelTest do
     assert_receive {:EXIT, ^pid, :shutdown}
   end
 
+  @tag :capture_log
   test "pushes and broadcast messages" do
     socket = subscribe_and_join!(socket(), Channel, "foo:ok")
     refute_broadcast "broadcast", _params
@@ -270,6 +279,7 @@ defmodule Phoenix.Test.ChannelTest do
     assert_broadcast "broadcast", %{"foo" => "bar"}
   end
 
+  @tag :capture_log
   test "connects and joins topics directly" do
     :error = connect(UserSocket, %{"reject" => true})
     {:ok, socket} = connect(UserSocket, %{})
@@ -286,12 +296,14 @@ defmodule Phoenix.Test.ChannelTest do
 
   ## handle_out
 
+  @tag :capture_log
   test "push broadcasts by default" do
     socket = subscribe_and_join!(socket(), Channel, "foo:ok")
     broadcast_from! socket, "default", %{"foo" => "bar"}
     assert_push "default", %{"foo" => "bar"}
   end
 
+  @tag :capture_log
   test "handles broadcasts and stops" do
     Process.flag(:trap_exit, true)
     {:ok, _, socket} = subscribe_and_join(socket(), Channel, "foo:ok")
@@ -303,6 +315,7 @@ defmodule Phoenix.Test.ChannelTest do
 
   ## handle_info
 
+  @tag :capture_log
   test "handles messages and stops" do
     Process.flag(:trap_exit, true)
     socket = subscribe_and_join!(socket(), Channel, "foo:ok")
@@ -312,12 +325,14 @@ defmodule Phoenix.Test.ChannelTest do
     assert_receive {:EXIT, ^pid, :shutdown}
   end
 
+  @tag :capture_log
   test "handles messages and pushes" do
     socket = subscribe_and_join!(socket(), Channel, "foo:ok")
     send socket.channel_pid, :push
     assert_push "info", %{"reason" => "push"}
   end
 
+  @tag :capture_log
   test "handles messages and broadcasts" do
     socket = subscribe_and_join!(socket(), Channel, "foo:ok")
     send socket.channel_pid, :broadcast
@@ -326,6 +341,7 @@ defmodule Phoenix.Test.ChannelTest do
 
   ## terminate
 
+  @tag :capture_log
   test "leaves the channel" do
     Process.flag(:trap_exit, true)
     {:ok, _, socket} = join(socket(), Channel, "foo:ok")
@@ -340,6 +356,7 @@ defmodule Phoenix.Test.ChannelTest do
     _ = leave(socket)
   end
 
+  @tag :capture_log
   test "closes the channel" do
     Process.flag(:trap_exit, true)
     {:ok, _, socket} = join(socket(), Channel, "foo:ok")
@@ -353,6 +370,7 @@ defmodule Phoenix.Test.ChannelTest do
     _ = close(socket)
   end
 
+  @tag :capture_log
   test "kills the channel when we reach timeout on close" do
     Process.flag(:trap_exit, true)
     {:ok, _, socket} = join(socket(), Channel, "foo:timeout")
@@ -378,6 +396,7 @@ defmodule Phoenix.Test.ChannelTest do
       {:error, :cant}
   end
 
+  @tag :capture_log
   test "subscribe and subscribe" do
     socket = subscribe_and_join!(socket(), Channel, "foo:ok")
     socket.endpoint.broadcast!("another:topic", "event", %{})
